+++ include/tictoc.h
@@ -41,6 +41,13 @@ public:
         }
     }
 
+    double toc()
+    {
+        end = std::chrono::system_clock::now();
+        std::chrono::duration<double> elapsed_seconds = end - start;
+        return elapsed_seconds.count() * 1000;
+    }
+
 private:  
     std::chrono::time_point<std::chrono::system_clock> start, end;
     bool disp_ = false;