+++ include/utility.h
@@ -4,6 +4,7 @@
 #define PCL_NO_PRECOMPILE 
 // <!-- liorf_yjz_lucky_boy -->
 #include <rclcpp/rclcpp.hpp>
+#include "liorf_diagnostics.h"
 
 #include <std_msgs/msg/header.hpp>
 #include <std_msgs/msg/string.hpp>
@@ -90,6 +91,8 @@ inline std::string TranslationPredictionSourceToString(TranslationPredictionSour
 class ParamServer : public rclcpp::Node
 {
 public:
+    std::shared_ptr<LiorfDiagnostics> diagnostics;
+
     string history_policy;
     string reliability_policy;